+++ EAMIS.MicroSrvc/EAMIS.Core/Domain/Entities/EAMISDELIVERYRECEIPTDETAILS.cs
@@ -12,7 +12,7 @@ namespace EAMIS.Core.Domain.Entities
         public int QTY_DELIVERED { get; set; }
         public int QTY_REJECTED { get; set;}
         public int QTY_RECEIVED { get; set; }
-        public int UNIT_COST { get; set; }
+        public decimal UNIT_COST { get; set; }
         public int SERIAL_LOT { get; set; }
         public string UNIT_OF_MEASUREMENT { get; set; }
         //public DateTime WARRANTY_EXPIRY_DATE { get; set; }
